requirements.txt
asgiref==3.9.1
atomicwrites==1.4.1
attrs==25.3.0
black==25.1.0
click==8.2.1
colorama==0.4.6
Django==5.2.4
djangorestframework==3.16.0
iniconfig==2.1.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
py==1.11.0
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
ruff==0.12.1
sqlparse==0.5.3
toml==0.10.2
tzdata==2025.2
